refactored routes functions into helper functions
